Migrated psar indicator to the class pattern.
