+++ js/Stock/Indicators/PSAR/PSARIndicator.js
@@ -10,8 +10,23 @@
  *
  * */
 'use strict';
-import BaseSeries from '../../Core/Series/Series.js';
-// im port './SMAIndicator.js';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../../Core/Series/Series.js';
+var SMAIndicator = BaseSeries.seriesTypes.sma;
+import U from '../../../Core/Utilities.js';
+var merge = U.merge, extend = U.extend;
 /* eslint-disable require-jsdoc */
 // Utils:
 function toFixed(a, n) {
@@ -84,6 +99,11 @@ function getPSAR(pdir, sDir, PSAR, pACCMulti, sLow, pLow, pHigh, sHigh, pEP) {
     return pEP;
 }
 /* eslint-enable require-jsdoc */
+/* *
+ *
+ * Class
+ *
+ * */
 /**
  * The Parabolic SAR series type.
  *
@@ -93,74 +113,26 @@ function getPSAR(pdir, sDir, PSAR, pACCMulti, sLow, pLow, pHigh, sHigh, pEP) {
  *
  * @augments Highcharts.Series
  */
-BaseSeries.seriesType('psar', 'sma', 
-/**
- * Parabolic SAR. This series requires `linkedTo`
- * option to be set and should be loaded
- * after `stock/indicators/indicators.js` file.
- *
- * @sample stock/indicators/psar
- *         Parabolic SAR Indicator
- *
- * @extends      plotOptions.sma
- * @since        6.0.0
- * @product      highstock
- * @requires     stock/indicators/indicators
- * @requires     stock/indicators/psar
- * @optionparent plotOptions.psar
- */
-{
-    lineWidth: 0,
-    marker: {
-        enabled: true
-    },
-    states: {
-        hover: {
-            lineWidthPlus: 0
-        }
-    },
-    /**
-     * @excluding period
-     */
-    params: {
-        /**
-         * The initial value for acceleration factor.
-         * Acceleration factor is starting with this value
-         * and increases by specified increment each time
-         * the extreme point makes a new high.
-         * AF can reach a maximum of maxAccelerationFactor,
-         * no matter how long the uptrend extends.
-         */
-        initialAccelerationFactor: 0.02,
-        /**
-         * The Maximum value for acceleration factor.
-         * AF can reach a maximum of maxAccelerationFactor,
-         * no matter how long the uptrend extends.
-         */
-        maxAccelerationFactor: 0.2,
-        /**
-         * Acceleration factor increases by increment each time
-         * the extreme point makes a new high.
+var PSARIndicator = /** @class */ (function (_super) {
+    __extends(PSARIndicator, _super);
+    function PSARIndicator() {
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        /* *
          *
-         * @since 6.0.0
-         */
-        increment: 0.02,
-        /**
-         * Index from which PSAR is starting calculation
+         *  Properties
          *
-         * @since 6.0.0
-         */
-        index: 2,
-        /**
-         * Number of maximum decimals that are used in PSAR calculations.
-         *
-         * @since 6.0.0
-         */
-        decimals: 4
+         * */
+        _this.data = void 0;
+        _this.points = void 0;
+        _this.options = void 0;
+        return _this;
     }
-}, {
-    nameComponents: false,
-    getValues: function (series, params) {
+    /* *
+     *
+     *  Functions
+     *
+     * */
+    PSARIndicator.prototype.getValues = function (series, params) {
         var xVal = series.xData, yVal = series.yData, 
         // Extreme point is the lowest low for falling and highest high
         // for rising psar - and we are starting with falling
@@ -214,8 +186,84 @@ BaseSeries.seriesType('psar', 'sma',
             xData: xData,
             yData: yData
         };
-    }
+    };
+    /**
+     * Parabolic SAR. This series requires `linkedTo`
+     * option to be set and should be loaded
+     * after `stock/indicators/indicators.js` file.
+     *
+     * @sample stock/indicators/psar
+     *         Parabolic SAR Indicator
+     *
+     * @extends      plotOptions.sma
+     * @since        6.0.0
+     * @product      highstock
+     * @requires     stock/indicators/indicators
+     * @requires     stock/indicators/psar
+     * @optionparent plotOptions.psar
+     */
+    PSARIndicator.defaultOptions = merge(SMAIndicator.defaultOptions, {
+        lineWidth: 0,
+        marker: {
+            enabled: true
+        },
+        states: {
+            hover: {
+                lineWidthPlus: 0
+            }
+        },
+        /**
+         * @excluding period
+         */
+        params: {
+            /**
+             * The initial value for acceleration factor.
+             * Acceleration factor is starting with this value
+             * and increases by specified increment each time
+             * the extreme point makes a new high.
+             * AF can reach a maximum of maxAccelerationFactor,
+             * no matter how long the uptrend extends.
+             */
+            initialAccelerationFactor: 0.02,
+            /**
+             * The Maximum value for acceleration factor.
+             * AF can reach a maximum of maxAccelerationFactor,
+             * no matter how long the uptrend extends.
+             */
+            maxAccelerationFactor: 0.2,
+            /**
+             * Acceleration factor increases by increment each time
+             * the extreme point makes a new high.
+             *
+             * @since 6.0.0
+             */
+            increment: 0.02,
+            /**
+             * Index from which PSAR is starting calculation
+             *
+             * @since 6.0.0
+             */
+            index: 2,
+            /**
+             * Number of maximum decimals that are used in PSAR calculations.
+             *
+             * @since 6.0.0
+             */
+            decimals: 4
+        }
+    });
+    return PSARIndicator;
+}(SMAIndicator));
+extend(PSARIndicator.prototype, {
+    nameComponents: false
 });
+BaseSeries.registerSeriesType('psar', PSARIndicator);
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default PSARIndicator;
 /**
  * A `PSAR` series. If the [type](#series.psar.type) option is not specified, it
  * is inherited from [chart.type](#chart.type).
