+++ js/masters/indicators/psar.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/PSARIndicator.js';
+import '../../Stock/Indicators/PSAR/PSARIndicator.js';
