+++ ts/Stock/Indicators/PSAR/PSARPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type PSARIndicator from './PSARIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class PSARPoint extends SMAPoint {
+    public series: PSARIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default PSARPoint;
